moved SettingsForm.tsx
moved SettingsFormButton,tsx

moved SettingsFormFields.tsx

moved SettingsFormFieldContent.tsx

moved SettingsFormImageFields.tsx

moved them to form folder
